+++ infra-market-server/src/main/kotlin/io/infra/market/service/UserService.kt
@@ -49,8 +49,8 @@ class UserService(
         val result = PageResultDto(
             records = userDtos,
             total = page.totalRow,
-            current = page.pageNumber.toInt(),
-            size = page.pageSize.toInt()
+            current = page.pageNumber,
+            size = page.pageSize
         )
         
         return ApiResponse.success(result)
